Feature: Reorganize and simplify again
Change the directory structure to contain a directory for each image containing a Dockerfile.
The tag name of the image is the name of the directory.

Add an EOL directory to which end of life image directories can be moved when they reach EOL.

Makefile can automatically deal with directory hierarchies. It will build based on the files
or directories in the current directory and if it is a directory containing a Makefile it
will issue a make command on that directory. The Makefile in subdirectories can/should
contain one line: include ../Makefile

Using this approach one will still be able to make EOL images by cd'ing to the EOL directory
(or a subdirectory if reorganzed again) and issue make [tag-dir-of-image].